pax-app/Wiki#184 Change ChatAddressBottomSheet to statefull and fetch user's addresses in its initState()
